requirements/base.txt
beautifulsoup4==4.4.1
Django==1.9.2
django-appconf==1.0.1
django-ckeditor==5.0.3
django-compressor==2.0
django-debug-toolbar==1.4
django-extensions==1.6.1
django-filebrowser-no-grappelli==3.6.2
django-geoposition==0.2.2
django-modelcluster==1.1
django-taggit==0.18.0
django-treebeard==3.0
djangorestframework==3.3.2
html5lib==0.9999999
Pillow==3.1.1
psycopg2==2.6.1
pytz==2015.7
rcssmin==1.0.6
rjsmin==1.0.12
singlemodeladmin==0.4
six==1.10.0
sqlparse==0.1.18
Unidecode==0.4.19
Unipath==1.1
django-uuslug==1.1.7
wheel==0.24.0
Willow==0.2.2